Moved readAddresses and processID to readmem.cpp. Cleaned some of the headers and moved more lua commands to runAutoSplitter
